[ROCm] Add HIP/ROCm support for AMD GPUs
Add support for building and running BaSpaCho on AMD GPUs via HIP/ROCm. The port
uses a minimal-footprint approach: a cuda_to_hip.h compat header aliases the CUDA
APIs to their HIP equivalents while keeping the source files in CUDA spelling, and
CMake marks the .cu files LANGUAGE HIP when building with -DUSE_HIP=ON. The default
CUDA build is unchanged. The README documents the HIP/ROCm build alongside the CUDA
build.

Library substitutions (behind USE_HIP):
- cuBLAS -> hipBLAS (GEMM, TRSM, SYMM, batched variants)
- cuSOLVER -> hipSOLVER (potrf, potrfBatched)
- cuSPARSE -> hipSPARSE

It also carries the Windows/amdclang++ build fixes needed to compile the GPU target
on Windows, all behind _WIN32 guards so the Linux build is byte-identical: a
typeid-name fallback where abi::__cxa_demangle is libstdc++-only, <malloc.h> for
alloca, localtime_s for localtime_r, and a few MSVC-ABI type and overload fixes.

This work was authored with the assistance of Claude, an AI assistant by Anthropic.

Test Plan:

Built with -DUSE_HIP=ON and ran the ctest suite on three AMD architectures:

```
cmake -S . -B build -DCMAKE_BUILD_TYPE=Release -DUSE_HIP=ON \
  -DCMAKE_HIP_ARCHITECTURES=<arch> -DBASPACHO_USE_BLAS=ON
cmake --build build -- -j$(nproc)
HIP_VISIBLE_DEVICES=0 ctest --test-dir build --output-on-failure
```

- gfx90a (Instinct MI250X, CDNA2), Linux ROCm 7.2.1: 124/125
- gfx1100 (Radeon Pro W7800, RDNA3), Linux ROCm 7.2.1: 124/125
- gfx1201 (Radeon RX 9070 XT, RDNA4), Windows ROCm 7.14: 124/125

The single failure (BatchedCudaFactor.CoalescedFactor_Many_float) is a float32 FMA
rounding tolerance overshoot (5.06e-05 vs the 5.0e-05 threshold) identical across all
GPU architectures including NVIDIA, not a ROCm-specific issue. The CPU build is unaffected.

+++ baspacho/baspacho/MatOpsCpuBase.h
@@ -211,8 +211,8 @@ struct CpuBaseNumericCtx : NumericCtx<T> {
 
   static inline void stridedMatSub(T* dst, int64_t dstStride, const T* src, int64_t srcStride,
                                    int64_t rSize, int64_t cSize) {
-    for (uint j = 0; j < rSize; j++) {
-      for (uint i = 0; i < cSize; i++) {
+    for (unsigned int j = 0; j < (unsigned int)rSize; j++) {
+      for (unsigned int i = 0; i < (unsigned int)cSize; i++) {
         dst[i] -= src[i];
       }
       dst += dstStride;
